moved functions to own module
